backend/requirements.txt
prisma
python-dotenv
fastapi
uvicorn
pydantic
supabase
google-genai
mcp
httpx
python-multipart
